refactor routes and models
