+++ backend.py
@@ -1,9 +1,4 @@
-from flask import Flask, request, jsonify, render_template, redirect, url_for, flash, session
-from flask_sqlalchemy import SQLAlchemy
-from flask_login import LoginManager, UserMixin, login_user, logout_user, login_required, current_user
-from werkzeug.security import generate_password_hash, check_password_hash
-import jwt
-import datetime
+from flask import Flask, request, jsonify, render_template, redirect, url_for, flash, session, Blueprint
 from functools import wraps
 import requests 
 import yfinance as yf 
@@ -11,12 +6,8 @@ import wikipedia
 from google import genai 
 from dotenv import load_dotenv 
 import os
-import secrets
-import authenticator
 from alert_system.scheduler import start_scheduler, alerts
-from flask_session import Session
-from flask_cors import CORS
-
+from utils import login_required
 # Load environment variables from .env file
 load_dotenv()
 
@@ -73,30 +64,16 @@ try:
 except Exception as e:
     print(f"Error initializing Gemini client: {e}")
     # We'll handle this in the query_gemini_llm function
-app = Flask(__name__, static_folder="static", template_folder="templates") 
-CORS(app)  # Enable CORS for all routes
 
+#initialize blueprint
+backend = Blueprint('backend', __name__, url_prefix='/service')
 
-#INITIAILIZE APP
-app = Flask(__name__, static_folder="static", template_folder="templates")
-
-# Configuration
-app.config['SECRET_KEY'] = secrets.token_hex(32)  # Change this in production!
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///stockmind.db'
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-app.config['JWT_SECRET_KEY'] = secrets.token_hex(32)  # For API tokens
-app.config['SESSION_TYPE'] = 'filesystem' #using server side session cookies - filesystem
-
-# Initialize Flask extensions
-db = SQLAlchemy(app)
-Session(app)
-
-@app.route('/alert_form')
+@backend.route('/alert_form')
 def alert_form():
     return render_template('alert_form.html')
-@app.route('/alerts')
+@backend.route('/alerts')
 
-@app.route('/create_alert', methods=['POST'])
+@backend.route('/create_alert', methods=['POST'])
 def create_alert():
     data = request.form
     alerts.append({
@@ -112,40 +89,7 @@ def create_alert():
 
 start_scheduler()
 
-# User model
-class User(db.Model):
-    id = db.Column(db.Integer, primary_key = True)
-    username = db.Column(db.String(50), nullable = False)
-    email = db.Column(db.String(120), unique=True, nullable = False)
-    password_hash = db.Column(db.String(200), nullable = False)
-    def set_passsword(self, passw):
-        self.password_hash = generate_password_hash(passw)
-    def check_password(self, passw):
-        return check_password_hash(self.password_hash, passw)
-    def get_passw_hash(self):
-        return self.password_hash
-
-# JWT token required decorator for API routes
-def token_required(f):
-    @wraps(f)
-    def decorated(*args, **kwargs):
-        token = None
-        
-        if 'Authorization' in request.headers:
-            token = request.headers['Authorization'].split(" ")[1]
-            
-        if not token:
-            return jsonify({'message': 'Token is missing!'}), 401
-            
-        try:
-            data = jwt.decode(token, app.config['JWT_SECRET_KEY'], algorithms=["HS256"])
-            current_user = User.query.filter_by(id=data['user_id']).first()
-        except:
-            return jsonify({'message': 'Token is invalid!'}), 401
-            
-        return f(current_user, *args, **kwargs)
-        
-    return decorated
+#helper functions 
 
 def fetch_wikipedia_summary(company_name): 
     try: 
@@ -158,7 +102,6 @@ def fetch_wikipedia_summary(company_name):
         return None, f"Error fetching Wikipedia summary: {str(e)}" 
     return None, "No Wikipedia page found for the given company." 
  
-
 def fetch_stock_price(ticker): 
     try: 
         # Set a timeout for the request
@@ -190,7 +133,6 @@ def fetch_stock_price(ticker):
         base_price = 100.0
         stock_prices = [round(base_price + random.uniform(-10, 10), 2) for _ in range(90)]
         return stock_prices, time_labels
-
 
 def get_ticker_from_alpha_vantage(company_name): 
     # Check if company is in our cache first
@@ -395,95 +337,10 @@ def query_gemini_llm(description):
             }
         ]
  
-def userAuthenticate():
-    '''use inside route functions to block logged out user'''
-    if "username" in session:
-        return True
-    return False
 
-# Authentication Routes
-@app.route("/login", methods = ['POST'])
-def login():
-    email = request.form['email']
-    password = request.form['password']
-    if(not email or not password):
-        return render_template("access-account.html", error = "Invalid Information, Please try again")
-    user = User.query.filter_by(email = email).first()
-    if(user and user.check_password(password)):
-        username = user.username
-        session["username"] = username
-        return redirect(url_for("home"))
-    else:
-        return render_template("access-account.html", error = "Invalid Information")
-        
-@app.route("/register", methods = ['POST'])
-def register():
-    username = request.form['username']
-    email = request.form['email']
-    password = request.form['password']
-    if(not username or not email or not password):
-        return render_template("access-account.html", error = "Invalid Information, Please try again")
-    user = User.query.filter_by(email = email).first()
-    if user:
-        return render_template("access-account.html", error="Account Already Exist")
-    else:
-        new_user = User(username = username, email = email)
-        new_user.set_passsword(password)
-        session['username'] = username
-        session['password'] = new_user.get_passw_hash()
-        session['email'] = email
-        return redirect(url_for('auth'))
-    
-@app.route('/api/auth')
-def auth():
-    username = session['username']
-    email = session['email']
-    try:
-        otp =  authenticator.generateOTP(username=username, usermail=email)
-        session["otp"] = otp
-    except:
-        return render_template("access-account.html", error = "Invalid email address")
-    return render_template("access-account.html", otp = True)
-
-@app.route('/api/verify', methods = ['POST'])
-def verify():
-    inp = request.form['userOTP']
-    username = session["username"]
-    password= session["password"]
-    email = session["email"]
-    otp = session["otp"]
-    session.pop("password",None)
-    session.pop("email",None)
-    session.pop("otp",None)
-    authSuccess = authenticator.verifyOTP(otp, inp)
-    if(authSuccess):
-        newUser = User(username = username , email = email, password_hash = password)
-        #registering the user in database
-        db.session.add(newUser)
-        db.session.commit()
-        return redirect(url_for('home'))
-    else:
-        session.pop("username",None)
-        return render_template("FRONT.html", error = "❌ Invalid OTP")
-    
-@app.route('/logout')
-def logout():
-    session.pop("username", None)
-    return redirect(url_for('home'))
-
-@app.route('/access-account')
-def accessAccount():
-    return render_template("access-account.html")
-
-# Protect existing routes
-@app.route("/")
-def home():
-    return render_template("FRONT.html")
-
-@app.route("/analyze_company", methods=["GET"])
+@backend.route("/analyze_company", methods=["GET"])
+@login_required
 def analyze_company():
-    if not userAuthenticate():
-        return render_template("FRONT.html", error = "Please Sign In to continue")
     company_name = request.args.get("company_name")
     if not company_name:
         return jsonify(success=False, error="No company name provided.")
@@ -493,7 +350,7 @@ def analyze_company():
         return jsonify(success=False, error="Could not find company description.")
 
     ticker = get_ticker_from_alpha_vantage(company_name)
-    if not ticker:
+    if not ticker:  
         return jsonify(success=False, error="Could not find ticker symbol.")
 
     stock_prices, time_labels = fetch_stock_price(ticker)
@@ -516,10 +373,3 @@ def analyze_company():
         competitors=competitors,
         top_competitors=top_competitors,
     )
-
-# Initialize database
-with app.app_context():
-    db.create_all()
-
-if __name__ == "__main__":
-    app.run(debug=True)
