+++ templates/access-account.html
@@ -111,13 +111,13 @@
     <div class="tab" id="registerTab">Sign Up</div>
   </div>
   <div class="forms">
-    <form id="loginForm" method="POST" action="/login">
+    <form id="loginForm" method="POST" action="/account/login">
         <input type="email" name="email" placeholder="Email" required/>
         <input type="password" name="password" placeholder="Password" required/>
         <button type="submit" class="login-btn">Login</button>
     </form>
 
-    <form id="registerForm" method="POST" action="/register">
+    <form id="registerForm" method="POST" action="/account/register">
       <input type="text" name="username" placeholder="Username" required />
       <input type="email" name="email" placeholder="Email" required />
       <input type="password" name="password" placeholder="Password (at least 8 characters)" required minlength="8"/>
